Modify main.js for libRuler
Make functions static so they can be called directly from Ruler.measure without instantiating the PathfindingRuler instance (see libruler.js wrap of measure):
- convertLocationToGridspace
- convertGridspaceToLocation 
- hitsWall
- rebuildGrid
- isInList
- isValidNode
- getNeighbors
- pruneWaypoints

update findPath to rely on the above static functions. For libRuler, findPath will just update waypoints, so avoid removing or measuring the ruler when libRuler is active. This allows us to use findPath as a method in Ruler or as a method in PathfindingRuler, avoiding duplicate code.